Permit running from Docker
In theory, this prevents the need to install ruby and the gems on the
local system. They should all be installed in the container.

Note that the current scripts assume the app folder will be mounted
to the container, but there is a line in the Dockerfile commented out
that re-enables a manual copy of the files to keep the server
completely separate from the local copy. The advantage of the bind is
that local changes to the code will be mirrored to the Docker
deployment without any further need to restart.